Refactored ShareCommentPanel out of ChatWindow
